+++ src/josepy/jwa.py
@@ -5,13 +5,16 @@ https://tools.ietf.org/html/draft-ietf-jose-json-web-algorithms-40
 """
 import abc
 import logging
-from typing import Dict, Type
+from typing import Dict, Type, Union
 
 import cryptography.exceptions
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives import hmac
-from cryptography.hazmat.primitives.asymmetric import padding, ec
+from cryptography.hazmat.primitives.asymmetric import ec, x25519, x448
+from cryptography.hazmat.primitives.asymmetric import ed25519
+from cryptography.hazmat.primitives.asymmetric import ed448
+from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography.hazmat.primitives.asymmetric.utils import decode_dss_signature
 from cryptography.hazmat.primitives.asymmetric.utils import encode_dss_signature
 
@@ -223,6 +226,36 @@ class _JWAEC(JWASignature):
             return True
 
 
+class _JWAOKP(JWASignature):
+    kty = jwk.JWKOKP
+
+    def __init__(self, name, hash_):
+        super().__init__(name)
+        self.hash = hash_()
+
+    def sign(self, key: Union[
+        ed25519.Ed25519PrivateKey,
+        ed448.Ed448PrivateKey,
+        x25519.X25519PrivateKey,
+        x448.X448PrivateKey,
+    ], msg: bytes):
+        return key.sign(msg)
+
+    def verify(self, key: Union[
+        ed25519.Ed25519PublicKey,
+        ed448.Ed448PublicKey,
+        x25519.X25519PrivateKey,
+        x448.X448PrivateKey,
+    ], msg: bytes, sig: bytes):
+        try:
+            key.verify(signature=sig, data=msg)
+        except cryptography.exceptions.InvalidSignature as error:
+            logger.debug(error, exc_info=True)
+            return False
+        else:
+            return True
+
+
 #: HMAC using SHA-256
 HS256 = JWASignature.register(_JWAHS('HS256', hashes.SHA256))
 #: HMAC using SHA-384
@@ -251,5 +284,12 @@ ES384 = JWASignature.register(_JWAEC('ES384', hashes.SHA384))
 #: ECDSA using P-521 and SHA-512
 ES512 = JWASignature.register(_JWAEC('ES512', hashes.SHA512))
 
-# Also implement RFC 8037, signing for OKP key type
-# hashes.BLAKE2b
+#: Ed25519 uses SHA512
+ES25519 = JWASignature.register(_JWAOKP('ES25519', hashes.SHA512))
+
+#: Ed448 uses SHA3/SHAKE256
+ES448 = JWASignature.register(_JWAOKP('ES448', hashes.SHAKE256))
+
+#: X25519
+
+#: X448