+++ src/josepy/jwa_test.py
@@ -3,6 +3,7 @@ import unittest
 from unittest import mock
 
 from josepy import errors, test_util
+from josepy.jwa import ES25519
 
 RSA256_KEY = test_util.load_rsa_private_key('rsa256_key.pem')
 RSA512_KEY = test_util.load_rsa_private_key('rsa512_key.pem')
@@ -228,6 +229,20 @@ class JWAECTest(unittest.TestCase):
             decode_patch.return_value = (0, 0)
             sig = ES512.sign(key.key, b"test")
             self.assertEqual(len(sig), 2 * 66)
+
+
+class JWAOKPTests(JWASignatureTest):
+    # look up the signature sizes in the RFC
+
+    def test_sign_no_private_part(self):
+        from josepy.jwa import ES25519
+        self.assertRaises(errors.Error, ES25519.sign, OKP_ED25519_KEY, b'foo')
+
+    def test_can_size_ed25519(self):
+        ES25519.sign(b'foo'), OKP_ED25519_KEY,
+
+    def test_signature_size(self):
+        pass
 
 
 if __name__ == '__main__':