+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 * Removed pytest-cache testing dependency.
 * Added support for Ed25519, Ed448, X25519 and X448 keys (see `RFC 8037 <https://tools.ietf.org/html/rfc8037>`_).
   These are also known as Bernstein curves.
+* Added support for signing with Ed25519, Ed448, X25519 and X448 keys
+  (see `RFC 8032 <https://datatracker.ietf.org/doc/html/rfc8032>`_).
 * Minimum requirement of ``cryptography`` is now 2.6+.
 
 1.8.0 (2021-03-15)